+++ src/main/java/pl/plajer/murdermystery/arena/ArenaRegistry.java
@@ -20,6 +20,7 @@ package pl.plajer.murdermystery.arena;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
@@ -43,6 +44,7 @@ public class ArenaRegistry {
 
   private static Main plugin = JavaPlugin.getPlugin(Main.class);
   private static List<Arena> arenas = new ArrayList<>();
+  private static int bungeeArena = -999;
 
   /**
    * Checks if player is in any arena
@@ -187,5 +189,16 @@ public class ArenaRegistry {
 
   public static List<Arena> getArenas() {
     return arenas;
+  }
+
+  public static void shuffleBungeeArena() {
+    bungeeArena = new Random().nextInt(arenas.size());
+  }
+
+  public static int getBungeeArena() {
+    if (bungeeArena == -999) {
+      bungeeArena = new Random().nextInt(arenas.size());
+    }
+    return bungeeArena;
   }
 }